+++ Demo/Harbeth-SwiftUI-Demo/Views/CurvesView.swift
@@ -33,7 +33,7 @@ struct CurvesView: View {
         NavigationView {
             VStack {
                 // 图像预览
-                HarbethView(image: inputImage, filters: [createFilter()]) {
+                DemoFilteredImage(image: inputImage, filters: [createFilter()]) {
                     $0.resizable()
                         .aspectRatio(contentMode: .fit)
                         .frame(maxWidth: .infinity, maxHeight: 400)